requirements.txt
python-telegram-bot==21.3
APScheduler==3.10.4
python-dotenv==1.0.0